Fix: connection issue on Win and Mac
(cherry picked from commit 2a427cfe5fea3f52a9a2f63b870d16277dc073ad)
(cherry picked from commit 93d5d0d67606161fbd50a062386e0ec7bbee6c54)

+++ src/net.h
@@ -623,7 +623,7 @@ public:
     {
         try {
             BeginMessage(pszCommand);
-            ssSend << a1 << (unsigned long)a2 << a3 << a4 << a5 << a6 << a7 << a8 << a9;
+            ssSend << a1 << a2 << a3 << a4 << a5 << a6 << a7 << a8 << a9;
             EndMessage();
         } catch (...) {
             AbortMessage();
